slsd

+++ models.py
@@ -16,6 +16,6 @@ class Product(Base):
     product_id = Column(Integer, primary_key=True)
     product_name = Column(String)
     product_quantity = Column(Integer)
-    product_price = Column(Float)
+    product_price = Column(Integer)
     date_updated = Column(Date)
     brand_id = Column(Integer, ForeignKey('brands.brand_id'))
